Fixed recording and resource manager

+++ core/src/commonMain/kotlin/ResourceManager.kt
@@ -1,7 +1,7 @@
 package pl.lemanski.pandaloop
 
 expect object ResourceManager {
-    fun saveAudioFile(path: String, buffer: FloatArray)
+    fun saveAudioFile(path: String, buffer: ByteArray)
 
-    fun loadAudioFile(path: String): FloatArray
+    fun loadAudioFile(path: String): ByteArray
 }
